+++ src/containers/ConnectForTableContent.js
@@ -1,0 +1,30 @@
+import { connect } from 'react-redux';
+import TableContent from '../components/TableContent';
+import rowsSortAndFilterSelector from '../store/selectors';
+import * as actions from '../store/actions';
+
+const mapStateToProps = state => {
+  const {
+    columnOrder = [],
+    defaultFixedRowsColumnWidth,
+    defaultRowHeight,
+    defaultHeaderRowHeight,
+    rows = [],
+    fixedColumnsCount,
+  } = state.tableData;
+  return {
+    columnOrder,
+    defaultFixedRowsColumnWidth,
+    filteredRowIndexes: rowsSortAndFilterSelector(state),
+    defaultRowHeight,
+    defaultHeaderRowHeight,
+    rows,
+    fixedColumnsCount,
+  };
+};
+
+const mapDispatchToProps = dispatch => ({
+  onRemoveSelectedRows: selectedRows => dispatch(actions.removeSelectedRows(selectedRows)),
+});
+
+export default connect(mapStateToProps, mapDispatchToProps)(TableContent);
